feat: add grouped blueprint cache
