+++ packages/application-generic/src/services/cache/interceptors/cached-entity.interceptor.ts
@@ -1,11 +1,16 @@
 import { Inject, Logger } from '@nestjs/common';
 
-import { CacheService } from '../cache.service';
+import { CacheService, CachingConfig } from '../cache.service';
 
 const LOG_CONTEXT = 'CachedEntityInterceptor';
-
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export function CachedEntity({ builder }: { builder: (...args) => string }) {
+export function CachedEntity({
+  builder,
+  options,
+}: {
+  builder: (...args) => string;
+  options?: CachingConfig;
+}) {
   const injectCache = Inject(CacheService);
 
   return (target: any, key: string, descriptor: any) => {
@@ -41,7 +46,7 @@ export function CachedEntity({ builder }: { builder: (...args) => string }) {
       const response = await originalMethod.apply(this, args);
 
       try {
-        await cacheService.set(cacheKey, JSON.stringify(response));
+        await cacheService.set(cacheKey, JSON.stringify(response), options);
       } catch (err) {
         // eslint-disable-next-line no-console
         Logger.error(
